Quarterly update for Azure IoT 2.10 Plugin
This commit consolidates work done for the Azure IoT 2.10 Plugin
release, based on TI's SimpleLink 2.30 SDKs (AZUREIOT-29) and
Microsoft's Azure IoT SDK LTS Oct 2018 build (AZUREIOT-28).

DPS support was added (AZUREIOT-26).  This required the addition of
the Parson library, as well as the socketio layer to the SimpleLink
PAL.  Notably:

- By default, Parson requires files support, but it was changed to a
  'lite' version (no file dependencies) to be friendlier to
  embedded devices.
- We can revisit (and ideally remove) the socketio dependency when
  Microsoft addresses issue #456:
  https://github.com/Azure/azure-iot-sdk-c/issues/456

Logging was added to the SDK and PAL debug libraries (AZUREIOT-24).

products.mak was made more intelligent for Plugin users by picking up
the installation locations of the various components if they were
previously set in the plugin or SDK (AZUREIOT-33).

We addressed the following items based on feedback from Microsoft
(AZUREIOT-41):

- Null argument checks in functions
- Use of macros from shared_util_options.h
- Additional error messages
- Redundant/Unnecessary assignments
- Better error handling in HTTPAPI_ExecuteRequest
- Replacement of tickcounter_sl.c with tickcounter_linux.c from Azure SDK
- Use mallocAndStrcpy_s() to simplify code
- Allocate contentBuf in HTTPAPI_ExecuteRequest from the heap to reduce
  stack space requirements
- Discard response body when responseContent buffer is NULL
- Add comment on receive buffer size in tlsio

Signed-off-by: Chris Ring <cring@ti.com>